Moved phone utilities to library
